Renamed PropertiesWidget files to match the class name
