+++ qtEditor/PropertiesWidget.h
